Port the four digit display to hooks
